starting with the scheduling of the repeated notifications
